Remove DOTFILES_ALREADY_RAN flag
Previously, dotfiles skipped initialization if
DOTFILES_ALREADY_RAN is set. This caused problems
with parent-child shell relationships, such as
shells embedded in IDEs and other similar things.

Re-running startup is mostly harmless; the thing
we really want to avoid is applying interactive
shell configuration in a noninteractive shell.

To support that, DOTFILES_ALREADY_RAN has been
removed; instead, functions that care about the
environment just explicitly check that PS1 is set
and, if when it isn’t, just return.