Fix bug on macOS using declare or local with -l
Using custom Bash 4.4.0 on macOS via homebrew appears unable to
use the declare or local builtins with the -l option to forcibly
convert any assigned value to lower case.

I’m not wholly certain where the bug lies — it could be a problem
with my local configuration, a bug in Bash, or simply a bug in
the macOS version of Bash that homebrew installs. Since it’s a
relatively obscure option, though, I’ve simply rewritten the two
places in the code using it to use tr instead.